+++ src/objects/impls/creep.rs
@@ -523,8 +523,8 @@ impl HasHits for Creep {
     }
 }
 
-impl MaybeHasNativeId for Creep {
-    fn try_native_id(&self) -> Option<JsString> {
+impl MaybeHasId<Creep> for Creep {
+    fn try_js_raw_id(&self) -> Option<JsString> {
         self.id_internal()
     }
 }