+++ src/objects/impls/deposit.rs
@@ -57,8 +57,8 @@ impl HasCooldown for Deposit {
     }
 }
 
-impl HasNativeId for Deposit {
-    fn native_id(&self) -> JsString {
+impl HasId<Deposit> for Deposit {
+    fn js_raw_id(&self) -> JsString {
         self.id_internal()
     }
 }