+++ src/objects/impls/ruin.rs
@@ -58,8 +58,8 @@ impl CanDecay for Ruin {
     }
 }
 
-impl HasNativeId for Ruin {
-    fn native_id(&self) -> JsString {
+impl HasId<Ruin> for Ruin {
+    fn js_raw_id(&self) -> JsString {
         Self::id_internal(self)
     }
 }