+++ src/objects/impls/symbol_decoder.rs
@@ -37,8 +37,8 @@ extern "C" {
     pub fn score_multiplier(this: &SymbolDecoder) -> u32;
 }
 
-impl HasNativeId for SymbolDecoder {
-    fn native_id(&self) -> JsString {
+impl HasId<SymbolDecoder> for SymbolDecoder {
+    fn js_raw_id(&self) -> JsString {
         Self::id_internal(self)
     }
 }